Hotel Bill
Suppose you went to a hotel. You ordered your meal and ate it. Now it’s time to pay the bill
for what you ate. There is a bill in front of you in which they mentioned the meal price (base
cost of a meal), tip percentage (the percentage of the meal price being added as tip), and tax
percentage (the percentage of the meal price being added as tax) for a meal. You need to
calculate the final price using a Java program. (Read input from Console using Scanner Class).