Prompting to temperature
One of my friend’s murali got a job offer in Kashmir and he is estimating the climate there
when packing his luggage. Based on the climate he needs to pack the specific clothing
accordingly for the conditions given below. Read temperature in centigrade and display a
suitable message according to temperature state below:
• Temp &lt; 0 then Freezing weather
• Temp 0-10 then Very Cold weather
• Temp 11-20 then Cold weather
• Temp 21-30 then Normal in Temp
• Temp 31-40 then It’s Hot
• Temp &gt;40 then It’s Very Hot
Help him to identify the climatic conditions. Now your task is to read