Electricity Bill
Chris and Dolly were discussing about their monthly bills to be paid in the due time and
identified that they are pending in paying the electrical bill and Dolly asks for the electrical
bill details. Here are the details given by Chris for the payment of bill. Generate electricity
bill based on numbers of units consumed
if units are &lt;=50 then rate per unit is 2RS.
if units are &gt;= 51 and &lt;= 100 then rate per unit is 3RS.
If units are&gt;= 101 and &lt;= 300 then rate per unit is 5 Rs and add extra 100Rs to the total bill.
If units are &gt;=301 and &lt;= 450 then rate per unit is 6RS and add extra 200 RS to the total bill.
if units are &gt;450 then per unit is 8RS and add extra 250 RS to the total bill.
Now your task is to calculate the amount to be pay.( Read units input from Console using
Scanner Class).