Operations on 2D Array
Develop a java program to that reads the number of rows, columns and reads the set of
elements in a 2D array from console. The menu driven program must perform the following
operations:
a) Sum of all elements
b) Print the data in matrix form
c) Print the elements of principal diagonal
d) Print the sum of elements in Principal diagonal.