C for C language
Raju is trying to install some programming languages software and He is going through the
given list. The list is as whenever a letter is given the corresponding language is displayed on the
screen for Raju’s selection to be installed. For example, if Raju enters ‘A’ or ‘a’ then he can select
“Ada” language likewise there many languages in which he can choose. Find out which language
Raju has chosen to install. (Hint: The user types in a letter of the alphabet and your program should
display the corresponding language or package available. Some sample input and output are given
below
Input Output
A or a Ada
C or c C Language
F or f Fortran
J or j Java
P or p Python
V or v Visual C++
Using the ‘switch’ statement to choose and display the appropriate message. Use the default label to
display a message if the input does not match any of the above letters.